Address Cosmos3 CUDA graph review feedback
