+++ vllm_omni/diffusion/models/cosmos3/cuda_graph.py
@@ -22,6 +22,12 @@ from vllm.logger import init_logger
 from vllm.platforms import current_platform
 
 from vllm_omni.diffusion.cuda_graph import DiffusionCUDAGraphConfig
+from vllm_omni.diffusion.models.cosmos3.utils import (
+    _clone_static,
+    _copy_inputs,
+    _key_for_value,
+    _return_output,
+)
 
 logger = init_logger(__name__)
 
@@ -82,6 +88,8 @@ class Cosmos3CUDAGraphManager:
                 "cache_size": len(self._cache),
                 "key_size": len(key),
             }
+            # Capture records the graph; replay fills static_output for this first request.
+            entry.graph.replay()
             return _return_output(entry.static_output, clone=self.config.clone_outputs)
 
         self._cache.move_to_end(key)
@@ -172,72 +180,3 @@ def _build_cached_gen_key(branch: str, kwargs: Mapping[str, Any]) -> tuple[Hasha
         if name in kwargs:
             key_parts.append((name, _key_for_value(kwargs[name])))
     return tuple(key_parts)
-
-
-def _key_for_value(value: Any) -> Hashable:
-    if value is None:
-        return ("none",)
-    if isinstance(value, torch.Tensor):
-        device = value.device
-        return (
-            "tensor",
-            tuple(value.shape),
-            str(value.dtype),
-            (device.type, device.index),
-            str(value.layout),
-            tuple(value.stride()),
-        )
-    if isinstance(value, (bool, int, float, str, bytes)):
-        return (type(value).__name__, value)
-    if isinstance(value, tuple):
-        return ("tuple", len(value), tuple(_key_for_value(item) for item in value))
-    if isinstance(value, list):
-        return ("list", len(value), tuple(_key_for_value(item) for item in value))
-    raise TypeError(f"unsupported key value {type(value)!r}")
-
-
-def _clone_static(value: Any) -> Any:
-    if isinstance(value, torch.Tensor):
-        return value.clone()
-    if isinstance(value, tuple):
-        return tuple(_clone_static(item) for item in value)
-    if isinstance(value, list):
-        return [_clone_static(item) for item in value]
-    if isinstance(value, Mapping):
-        return {key: _clone_static(item) for key, item in value.items()}
-    return value
-
-
-def _copy_inputs(static_value: Any, dynamic_value: Any) -> None:
-    if isinstance(static_value, torch.Tensor):
-        static_value.copy_(dynamic_value)
-        return
-    if isinstance(static_value, tuple):
-        for static_item, dynamic_item in zip(static_value, dynamic_value, strict=True):
-            _copy_inputs(static_item, dynamic_item)
-        return
-    if isinstance(static_value, list):
-        for static_item, dynamic_item in zip(static_value, dynamic_value, strict=True):
-            _copy_inputs(static_item, dynamic_item)
-        return
-    if isinstance(static_value, dict):
-        for key in static_value:
-            _copy_inputs(static_value[key], dynamic_value[key])
-
-
-def _return_output(output: Any, *, clone: bool) -> Any:
-    if not clone:
-        return output
-    return _clone_output(output)
-
-
-def _clone_output(value: Any) -> Any:
-    if isinstance(value, torch.Tensor):
-        return value.clone()
-    if isinstance(value, tuple):
-        return tuple(_clone_output(item) for item in value)
-    if isinstance(value, list):
-        return [_clone_output(item) for item in value]
-    if isinstance(value, Mapping):
-        return {key: _clone_output(item) for key, item in value.items()}
-    return value
